health_card_generator/requirements.txt
pillow==10.0.0
reportlab==3.6.12
qrcode==7.4.2
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
